Extract business logic as useBusinesses hook.#3
