+++ src/screens/SearchScrean.js
@@ -1,27 +1,11 @@
 import React, { useState, useEffect } from "react"
 import { View, Text, StyleSheet } from "react-native"
 import SearchBar from "../components/SearchBar"
-import yelp from "../services/httpServices"
+import useBusinesses from "../hooks/useBusinesses"
 
 const SearchScreen = () => {
   const [query, setQuery] = useState("")
-  const [results, setResults] = useState([])
-  const [errorMessage, setErrorMessage] = useState("")
-
-  useEffect(() => {
-    searchApi()
-  }, [])
-
-  const searchApi = async searchQuery => {
-    try {
-      const response = await yelp.get("/search", {
-        params: { limit: 50, term: searchQuery, location: "irvine" }
-      })
-      setResults(response.data.businesses)
-    } catch (err) {
-      setErrorMessage("Oops! Something went wrong")
-    }
-  }
+  const [searchApi, results, errorMessage] = useBusinesses()
 
   return (
     <View>
